feat: Implemented MunicipalityService

+++ analyze/src/main/java/io/natron/noscope360/analyze/model/dto/MunicipalityOverviewDto.java
@@ -1,5 +1,5 @@
 package io.natron.noscope360.analyze.model.dto;
 
 public record MunicipalityOverviewDto(int id, String name, String region, String canton, String district,
-                                      int avgQuantitativeRating, int avgQualitativeRating) {
+                                      double avgQuantitativeRating, double avgQualitativeRating) {
 }
